requirements.txt
# Core Dependencies
streamlit>=1.28.0
openai>=1.0.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
regex>=2023.0.0

# Optional but Recommended
colorama>=0.4.6  # For colored terminal output
requests>=2.31.0  # For HTTP requests if needed

pypandoc>=1.11



# Development Dependencies (Optional)
# Uncomment if you're developing or contributing
# pytest>=7.4.0  # For unit testing
# black>=23.0.0  # For code formatting
# flake8>=6.0.0  # For code linting
# mypy>=1.0.0  # For type checking